+++ backend/infrahub/graphql/manager.py
@@ -1005,8 +1005,16 @@ class GraphQLSchemaManager:
         if not top_level:
             filters["isnull"] = graphene.Boolean()
 
+        if schema.display_label:
+            display_label_schema = schema.get_attribute("display_label")
+            filters.update(
+                get_attribute_type(kind=display_label_schema.kind).get_graphql_filters(
+                    name="display_label", include_properties=False, include_isnull=True
+                )
+            )
+
         if schema.human_friendly_id and top_level:
-            # HFID filter limited to top level because we can't filter on HFID for relationships (yet)
+            # NOTE: this can loosen to allow filtering at a non-top level once IFC-2110 is implemented
             filters["hfid"] = graphene.List(graphene.String)
 
         for attr in schema.attributes: