+++ docs/sidebars.ts
@@ -57,6 +57,7 @@ const sidebars: SidebarsConfig = {
           items: [
             'guides/installation',
             'guides/production-deployment',
+            'guides/configuration-changes',
             'guides/database-backup',
             'guides/upgrade',
             'guides/repository',
@@ -420,6 +421,7 @@ const sidebars: SidebarsConfig = {
             slug: 'release-notes/infrahub',
           },
           items: [
+            'release-notes/infrahub/release-1_6_2',
             'release-notes/infrahub/release-1_6_1',
             'release-notes/infrahub/release-1_6_0',
             'release-notes/infrahub/release-1_5_3',