+++ frontend/app/src/entities/role-manager/api/getGlobalPermissions.ts
@@ -1,17 +1,8 @@
-import { gql } from "@apollo/client";
+import { graphql } from "gql.tada";
 
-export const GET_ROLE_MANAGEMENT_GLOBAL_PERMISSIONS = gql`
-  query GET_ROLE_MANAGEMENT_GLOBAL_PERMISSIONS(
-    $search: String
-    $offset: Int
-    $limit: Int
-  ) {
-    CoreGlobalPermission(
-      any__value: $search
-      partial_match: true
-      offset: $offset
-      limit: $limit
-    ) {
+export const GET_ROLE_MANAGEMENT_GLOBAL_PERMISSIONS = graphql(`
+  query GET_ROLE_MANAGEMENT_GLOBAL_PERMISSIONS($search: String, $offset: Int, $limit: Int) {
+    CoreGlobalPermission(any__value: $search, partial_match: true, offset: $offset, limit: $limit) {
       count
       edges {
         node {
@@ -52,4 +43,4 @@ export const GET_ROLE_MANAGEMENT_GLOBAL_PERMISSIONS = gql`
       }
     }
   }
-`;
+`);